refactor: extract messaging into util
